# Improve storage and indexation normalization performance by using variadic function