<!--- (<3 Thanks for taking the time to contribute! You're awesome! <3) --->

<!--- (If you've never contributed to this repository before, please read https://github.com/akeneo/pim-community-dev/blob/master/.github/CONTRIBUTING.md) --->

**Description (for Contributor and Core Developer)**

### Problem

The normalization of PV is time consuming for the PHP process and the complexity is not linear.

As stated it in this article [No array_merge In Loops](https://github.com/dseguy/clearPHP/blob/master/rules/no-array_merge-in-loop.md), array_merge is inefficient in a loop.

```
$result = [];
foreach ($source as $list) {
  $result = array_merge($result, $list);
}
```

At first loop, array_merge copy the variable `$result` having a size of 1.
At second loop, array_merge copy the variable `$result` having a size of 2.
At third loop, array_merge copy the variable `$result` having a size of 3.... etc


The complexity is `*O(n^2)* (where n is the number of PV).  
Explanation: 1 + 2 + 3 + ....  is a suite. Formula: O(n*(n+1)/2) => O(n^2/2) => O(n^2)

### Solution

Use variadic function to not copy the `$result` array in each loop, but only one time.
The complexity is *O(n)* as seen in the graph.

And proof:
https://gist.github.com/ahocquard/e24346317511c359213bb79ba42652c2

![results_array_merge_10000](https://user-images.githubusercontent.com/1942439/46025582-9389e180-c0e9-11e8-93b6-4f8695968d88.png)

### How I saw that?

When I did a very simple POC to test:
- the limit of the DB
- by normalizing product values
- 1 transaction/product (not the real product of the PIM)
- 5000 PV

Before fix:
- 8 seconds for the saving part of the values in the DB
- 193 seconds for the normalization part (x24!)

After fix:
- 8 seconds for the saving part of the values in the DB
- 8 seconds for the normalization part


### What does it change in import?

I tested it quickly.
For now, almost nothing, as doing other tasks is far most costly due to the number of IO we are doing.
But it costs nothing and it will be a problem after optimisation of the PIM.
Normalisation of the product values is very important and have to be optimised the best we can (for reading and writing).


<!--- (What does this Pull Request do? reference the related issue?) --->

**Definition Of Done (for Core Developer only)**

| Q                                 | A
| --------------------------------- | ---
| Added Specs                       | Yes
| Added legacy Behats               | -
| Added acceptance tests            | -
| Added integration tests           | -
| Changelog updated                 | -
| Review and 2 GTM                  | Ok
| Micro Demo to the PO (Story only) | -
| Migration script                  | -
| Tech Doc                          | -

`Todo`: Pending / Work in progress
`OK`: Done / Validated
`-`: Not needed



Note: green CI, tests already failing on master